d3d11device: raise 'suspended' signal on DXGI_ERROR_DEVICE_REMOVED
This is the first step of the patch. When the error is cought the
GstD3D11Device object raises the new "suspended" signal. This allow
handling the error from outside: replace the catched GstContext by
the new one.
On the second step, GstD3D11Device will also manage to stop the
playback by returning NULL intead of the real handles, but this
will require to add changes to each plugin that uses GstD3D11Device,
otherwise it can crash accessing the NULL pointer.