d3d11d2d1: corrify the ID2D1Factory storage
Before we were storing one ID2D1Factory for all the instances
of the d2d1 element. But it seems to be dangerous due to
this article:
https://learn.microsoft.com/en-us/windows/win32/direct2d/multi-threaded-direct2d-apps#developing-thread-safe-direct2d-apps-with-minimal-direct3d-or-dxgi-calls

If we have one ID2D1Factory for all, then we can't protect it with the
device lock, so we decide to attach it to the GstD3D11Device object.
We assume that underneath some d2d1 calls there might be access to the
ID2D1Factory, and it might collide with other DXGI calls, so we are going
to make sure the factory is always protected with the device lock too.